requirements.txt
streamlit 
python-docx
re